Fix #365
Nested bullets are formatted incorrectly

Before: https://main--www--cmegroup.aem.page/education/courses/building-a-trade-plan/methodology-behind-your-trade-plan
After: https://issue-365-bulletSeparation-factoidCardVariant--www--cmegroup.aem.page/education/courses/building-a-trade-plan/methodology-behind-your-trade-plan

Part of the problem was with the nexted elements being auto-wrapped in p element which pulled in conflicting styling for what we needed.

This was a DA issue and a feature toggle was added to CME sites to maintain the original markup (not wrap with p element).

This fixed most of the issue. Only a small style override was required to fix the remaining issue.